Filter migration (#5555)
