+++ packages/core/src/filters/Filter.js
@@ -34,6 +34,9 @@ import defaultFragment from './defaultFilter.frag';
  * Developers can use normal coordinates of v3 and then allow filter to use partial Framebuffers,
  * bringing those extra uniforms into account.
  *
+ * Also be aware that we have changed default vertex shader, please consult
+ * {@link https://github.com/pixijs/pixi.js/wiki/v5-Creating-filters Wiki}.
+ *
  * ### Built-in Uniforms
  *
  * PixiJS viewport uses screen (CSS) coordinates, `(0, 0, renderer.screen.width, renderer.screen.height)`,
@@ -45,7 +48,7 @@ import defaultFragment from './defaultFilter.frag';
  * _Important note: as with all Framebuffers in PixiJS, both input and output are
  * premultiplied by alpha._
  *
- * By default, input Framebuffer space coordinates are passed to fragment shader with `vTextureCoord`.
+ * By default, input normalized coordinates are passed to fragment shader with `vTextureCoord`.
  * Use it to sample the input.
  *
  * ```
@@ -83,15 +86,15 @@ import defaultFragment from './defaultFilter.frag';
  *
  * **inputSize**
  *
- * Temporary Framebuffer is different, it can be either the size of screen, either power-of-two.
- * The `inputSize.xy` are size of temporary Framebuffer that holds input.
+ * Temporary framebuffer is different, it can be either the size of screen, either power-of-two.
+ * The `inputSize.xy` are size of temporary framebuffer that holds input.
  * The `inputSize.zw` is inverted, it's a shortcut to evade division inside the shader.
  *
  * Set `inputSize.xy = outputFrame.zw` for a fullscreen filter.
  *
- * To calculate input texture coordinate in 0-1 space, you have to map it to Framebuffer normalized space.
- * Multiply by `outputFrame.zw` to get pixel coordinate in part of Framebuffer.
- * Divide by `inputSize.xy` to get Framebuffer normalized space (input sampler space)
+ * To calculate input normalized coordinate, you have to map it to filter normalized space.
+ * Multiply by `outputFrame.zw` to get input coordinate.
+ * Divide by `inputSize.xy` to get input normalized coordinate.
  *
  * ```
  * vec2 filterTextureCoord( void )
@@ -209,11 +212,11 @@ export default class Filter extends Shader
      *        There are some useful properties in the currentState :
      *        target, filters, sourceFrame, destinationFrame, renderTarget, resolution
      */
-    apply(filterManager, input, output, clear, currentState, derp) // eslint-disable-line no-unused-vars
+    apply(filterManager, input, output, clear, currentState)
     {
         // do as you please!
 
-        filterManager.applyFilter(this, input, output, clear, currentState, derp);
+        filterManager.applyFilter(this, input, output, clear, currentState);
 
         // or just do a regular render..
     }
